GH-30: Implemented leave classroom functionality
