+++ code/js/public/css/style.css
@@ -69,7 +69,7 @@
   right: 0;
   bottom: 0;
   background-color: rgba(255,255,255,0.7);
-  z-index: 3;
+  z-index: 9999;
   backdrop-filter: blur(12px);
 }
